Proper bit-mask support (#198)

+++ include/roq/execution_instruction.hpp
@@ -17,7 +17,7 @@ namespace roq {
 //! Enumeration of execution types
 struct ROQ_PACKED ExecutionInstruction final {
   //! helper
-  enum type_t : uint8_t {
+  enum type_t : uint32_t {
     UNDEFINED = 0x0,
     PARTICIPATE_DO_NOT_INITIATE = 0x1,  //!< Cancel if order would have executed on placement (i.e. not as maker)
     CANCEL_IF_NOT_BEST = 0x2,           //!< Cancel if order can not be placed at best price
@@ -32,7 +32,7 @@ struct ROQ_PACKED ExecutionInstruction final {
   constexpr ExecutionInstruction(type_t type)  // NOLINT (allow implicit)
       : type_(type) {}
 
-  explicit constexpr ExecutionInstruction(uint8_t type) : type_(validate(type)) {}
+  explicit constexpr ExecutionInstruction(uint32_t type) : type_(validate(type)) {}
 
   constexpr operator type_t() const { return type_; }
 
@@ -58,7 +58,7 @@ struct ROQ_PACKED ExecutionInstruction final {
   constexpr operator std::string_view() const { return name(); }
 
  protected:
-  constexpr type_t validate(uint8_t type) {
+  constexpr type_t validate(uint32_t type) {
     auto result = static_cast<type_t>(type);
     switch (result) {
       case type_t::UNDEFINED:
@@ -84,7 +84,7 @@ struct std::is_enum<roq::ExecutionInstruction> : std::true_type {};
 
 template <>
 struct std::underlying_type<roq::ExecutionInstruction> {
-  using type = uint8_t;
+  using type = uint32_t;
 };
 
 template <>
